+++ packages/CHANGELOG.md
@@ -7,6 +7,11 @@
 
 * Add a `defaultAction` prop for `<PostAction />`, `<SaveAction />`, etc. components
 
+**`rev-ui-materialui`**
+
+* MUISelectField - store `null` instead of empty string when empty option is selected
+
+
 ### 0.18.0 - 6th June 2018
 
 * Build fix: clean lib/** before building